monitor/requirements.txt
solana-py==0.34.0
solders==0.21.0
anchorpy==0.20.1
scikit-learn==1.5.0
xgboost==2.0.3
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
python-telegram-bot==21.3
